Add REST API documentation
Add a doc directory with basic details on highlevel telemetry concepts
and an initial specification of the REST API.
