+++ doc/README.md
@@ -1,0 +1,8 @@
+# Telemetry Client Documentation
+
+The following topics are documented:
+* [Telemetry Blobs](telemetryblob.md)
+* [Telemetry Types](telemetrytype.md)
+* [Telemetry Tags](telemetrytag.md)
+* [Telemetry Timestamps](telemetrytimestamp.md)
+* [Telemetry Client REST API](api/)
