+++ doc/api/structs/telemetrydataitem.md
@@ -1,0 +1,37 @@
+# TelemetryDataItem Data Structure
+
+The TelemetryDataItem data structure consists of the following sections:
+
+* header - contains metadata about the telemetry report
+  * telemetryId - a UUID that differenciates this telemety data item
+    from other telemety data items generated by the same client
+  * telemetryTimeStamp - the UTC timestamp for when the telemetry data
+    item was generated, formatted in
+    [RFC3339nano](../../telemetrytimestamp.md) format
+  * telemetryType - the [Telemetry Type](../../telemetrytype.md)
+    associated with this telemetry submission.
+  * telemetryAnnotations - a possibly empty list of
+    [telemetry annotation tags](../../telemetrytag.md)
+* payload - a [JSON telemetry blob](../../telemetryblob.md)
+* footer - contains a checksum of the payload section
+
+***NOTE***: A telemetry data item doesn't specify a clientId in it's
+header; a clientId will be specified in the bundle that contains the
+this telemetry data item when it is reported to a telemetry server.
+
+```
+{
+	header {
+    telemetryId          string
+    telemetryTimeStamp   string($rfc3339nano)
+    telemetryType        string
+    telemetryAnnotations [
+      string...
+    ]
+  }
+	telemetryData       string($JSON)
+	footer {
+    checksum          string
+  }
+}
+```
