+++ doc/api/structs/telemetryreport.md
@@ -1,0 +1,39 @@
+# TelemetryReport Data Structure
+
+The TelemetryReport data structure consists of the following sections:
+
+* header - contains metadata about the telemetry report
+  * reportId - a UUID that differentiates this report from other reports
+    generated by the same client
+  * reportTimeStamp - the UTC timestamp for when the bundle was generated,
+    formatted in [RFC3339nano](../../telemetrytimestamp.md) format
+  * reportClientId - the clientId of the telemetry client generating this
+    report, as specified in the credentials returned when the telemetry
+    client successfully registered with the upstream telemetry server.
+  * reportAnnotations - a possibly empty list of
+    [telemetry annotation tags](../../telemetrytag.md)
+* payload - a list of one or more [TelemetryBundle](telemetrybundle.md) objects
+* footer - contains a checksum of the payload section
+
+***NOTE***: The [telemetry bundles](telemetrybundle.md) in a report payload
+may not all originate from the same telemetry client, e.g. they may have been
+relayed through one or more telemetry relays.
+
+```
+{
+	header {
+    reportId          string
+    reportTimeStamp   string($rfc3339nano)
+    reportClientId    integer($int64)
+    reportAnnotations [
+      string...
+    ]
+  }
+	telemetryBundles [
+    TelemetryBundle...
+  ]
+	footer {
+    checksum string
+  }
+}
+```
