+++ pkg/client/client.go
@@ -1,12 +1,10 @@
 package client
 
 import (
-	"bytes"
 	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"io/fs"
 	"log/slog"
 	"net/http"
@@ -16,7 +14,6 @@ import (
 
 	"github.com/SUSE/telemetry/pkg/config"
 	telemetrylib "github.com/SUSE/telemetry/pkg/lib"
-	"github.com/SUSE/telemetry/pkg/restapi"
 	"github.com/SUSE/telemetry/pkg/types"
 	"github.com/golang-jwt/jwt/v5"
 )
@@ -462,390 +459,6 @@ func unauthorizedError(resp *http.Response) (err error) {
 	}
 
 	return
-}
-
-func (tc *TelemetryClient) submitReportRetry(
-	report *telemetrylib.TelemetryReport,
-	maxTries int,
-	delay time.Duration,
-) (err error) {
-	// retry at most MaxTries times
-	for retry := maxTries; retry > 0; retry -= 1 {
-
-		// handle panic() calls as well as return
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					switch rType := r.(type) {
-					case string:
-						err = errors.New(rType)
-					case error:
-						err = rType
-					default:
-						err = fmt.Errorf("unexpected recovery type: %s", rType)
-					}
-				}
-			}()
-			err = tc.submitReportInternal(report)
-		}()
-
-		if err == nil {
-			break
-		}
-
-		switch {
-		// check if we need to register again?
-		case errors.Is(err, ErrRegistrationRequired):
-			slog.Info(
-				"Telemetry Client Registration Required",
-				slog.String("error", err.Error()),
-			)
-
-			// force a (re-)registration by deleting any existing
-			// client creds bundle
-			err = tc.deleteTelemetryAuth()
-			if err != nil {
-				slog.Warn(
-					"Failed to delete existing telemetry auth bundle",
-					slog.String("error", err.Error()),
-				)
-			}
-
-			// register the telemetry client
-			err = tc.Register()
-			if err != nil {
-				// if registration failed, for now don't re-try
-				return
-			}
-
-			slog.Info(
-				"Telemetry Client Registration Successful",
-			)
-
-		// check if we need to authenticate again?
-		case errors.Is(err, ErrAuthenticationRequired):
-			slog.Info(
-				"Telemetry Client Authentication Required",
-				slog.String("error", err.Error()),
-			)
-
-			// attempt to (re-)autenticate
-			err = tc.Authenticate()
-			if err != nil {
-				// if authentication failed, for now don't re-try
-				return
-			}
-
-			slog.Info(
-				"Telemetry Client Authentication Successful",
-			)
-
-		// TODO: handle server busy backoff and retry appropriately
-
-		default:
-			slog.Debug(
-				"Unhandled error",
-				slog.String("error", err.Error()),
-			)
-		}
-
-		// sleep between retries
-		if retry > 0 {
-			time.Sleep(delay)
-		}
-
-	}
-	return
-}
-
-func (tc *TelemetryClient) submitReport(report *telemetrylib.TelemetryReport) (err error) {
-	// TODO: make delay configurable, or possibly supplied by the request response
-	err = tc.submitReportRetry(report, 3, time.Duration(500*time.Millisecond))
-	return
-}
-
-func (tc *TelemetryClient) submitReportInternal(report *telemetrylib.TelemetryReport) (err error) {
-	// submit a telemetry report
-	var trReq restapi.TelemetryReportRequest
-	trReq.TelemetryReport = *report
-	reqBodyJSON, err := json.Marshal(&trReq)
-	if err != nil {
-		slog.Error("failed to JSON marshal trReq", slog.String("err", err.Error()))
-		return
-	}
-
-	reqUrl := tc.cfg.TelemetryBaseURL + "/report"
-	reqBuf := bytes.NewBuffer(reqBodyJSON)
-	req, err := http.NewRequest("POST", reqUrl, reqBuf)
-	if err != nil {
-		slog.Error("failed to create new HTTP request for telemetry report", slog.String("err", err.Error()))
-		return
-	}
-
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", "Bearer "+tc.auth.Token.String())
-	req.Header.Add("X-Telemetry-Client-Id", fmt.Sprintf("%d", tc.auth.ClientId))
-
-	httpClient := http.DefaultClient
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		slog.Error("failed HTTP POST telemetry report request", slog.String("err", err.Error()))
-		return
-	}
-	defer resp.Body.Close()
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		slog.Error(
-			"failed to read telemetry report response body",
-			slog.Int("StatusCode", resp.StatusCode),
-			slog.String("error", err.Error()),
-		)
-		return
-	}
-
-	switch resp.StatusCode {
-	case http.StatusOK:
-		// nothing to do
-	case http.StatusUnauthorized:
-		return unauthorizedError(resp)
-	default:
-		slog.Error(
-			"failed to submit report",
-			slog.Int("StatusCode", resp.StatusCode),
-			slog.String("respBody", string(respBody)),
-		)
-		return
-	}
-
-	var trResp restapi.TelemetryReportResponse
-	err = json.Unmarshal(respBody, &trResp)
-	if err != nil {
-		slog.Error("failed to JSON unmarshal telemetry report response body content", slog.String("err", err.Error()))
-		return
-	}
-
-	slog.Debug(
-		"successfully submitted report",
-		slog.String("report", report.Header.ReportId),
-		slog.String("processing", trResp.ProcessingInfo()),
-	)
-	return
-}
-
-// Authenticate is responsible for (re)authenticating an already registered
-// client with the server to ensure that it's auth token is up to date.
-func (tc *TelemetryClient) Authenticate() (err error) {
-	if err = tc.loadTelemetryAuth(); err != nil {
-		return fmt.Errorf(
-			"telemetry client (re-)authentication requires an existing "+
-				"client registration: %s",
-			err.Error(),
-		)
-	}
-
-	// get the instanceId, failing if it can't be retrieved
-	instId, err := tc.getInstanceId()
-	if err != nil {
-		return
-	}
-
-	// assemble the authentication request
-	caReq := restapi.ClientAuthenticationRequest{
-		ClientId:   tc.auth.ClientId,
-		InstIdHash: *instId.Hash("default"),
-	}
-
-	reqBodyJSON, err := json.Marshal(&caReq)
-	if err != nil {
-		slog.Error(
-			"failed to JSON marshal caReq",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	reqUrl := tc.cfg.TelemetryBaseURL + "/authenticate"
-	reqBuf := bytes.NewBuffer(reqBodyJSON)
-	req, err := http.NewRequest("POST", reqUrl, reqBuf)
-	if err != nil {
-		slog.Error(
-			"failed to create new HTTP request for client authentication",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	req.Header.Add("Content-Type", "application/json")
-
-	httpClient := http.DefaultClient
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		slog.Error(
-			"failed to HTTP POST client authentication request",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-	defer resp.Body.Close()
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		slog.Error(
-			"failed to read client authentication response body",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		err = fmt.Errorf("client authentication failed: %s", string(respBody))
-		return
-	}
-
-	var caResp restapi.ClientAuthenticationResponse
-	err = json.Unmarshal(respBody, &caResp)
-	if err != nil {
-		slog.Error(
-			"failed to JSON unmarshal client authentication response body content",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	tc.auth.ClientId = caResp.ClientId
-	tc.auth.Token = types.TelemetryAuthToken(caResp.AuthToken)
-	tc.auth.IssueDate, err = types.TimeStampFromString(caResp.IssueDate)
-	if err != nil {
-		slog.Error(
-			"failed to parse issueDate as a timestamp",
-			slog.String("issueDate", caResp.IssueDate),
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	err = tc.saveTelemetryAuth()
-	if err != nil {
-		slog.Error(
-			"failed to save TelemetryAuth",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	slog.Debug(
-		"successfully authenticated",
-		slog.Int64("clientId", tc.auth.ClientId),
-	)
-
-	return
-}
-
-func (tc *TelemetryClient) Register() (err error) {
-	// get the saved TelemetryAuth, returning success if found
-	err = tc.loadTelemetryAuth()
-	if err == nil {
-		slog.Debug("telemetry auth found, client already registered, skipping", slog.Int64("clientId", tc.auth.ClientId))
-		return
-	}
-
-	// get the instanceId, failing if it can't be retrieved
-	instId, err := tc.getInstanceId()
-	if err != nil {
-		return
-	}
-
-	// register the system as a client
-	var crReq restapi.ClientRegistrationRequest
-	crReq.ClientInstanceId = instId
-	reqBodyJSON, err := json.Marshal(&crReq)
-	if err != nil {
-		slog.Error(
-			"failed to JSON marshal crReq",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	reqUrl := tc.cfg.TelemetryBaseURL + "/register"
-	reqBuf := bytes.NewBuffer(reqBodyJSON)
-	req, err := http.NewRequest("POST", reqUrl, reqBuf)
-	if err != nil {
-		slog.Error(
-			"failed to create new HTTP request for client registration",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	req.Header.Add("Content-Type", "application/json")
-
-	httpClient := http.DefaultClient
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		slog.Error(
-			"failed to HTTP POST client registration request",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-	defer resp.Body.Close()
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		slog.Error(
-			"failed to read client registration response body",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	// TODO: Handle http.StatusConflict (409) as needing to regenerate instId
-	if resp.StatusCode != http.StatusOK {
-		err = fmt.Errorf("client registration failed: %s", string(respBody))
-		return
-	}
-
-	var crResp restapi.ClientRegistrationResponse
-	err = json.Unmarshal(respBody, &crResp)
-	if err != nil {
-		slog.Error(
-			"failed to JSON unmarshal client registration response body content",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	tc.auth.ClientId = crResp.ClientId
-	tc.auth.Token = types.TelemetryAuthToken(crResp.AuthToken)
-	tc.auth.IssueDate, err = types.TimeStampFromString(crResp.IssueDate)
-	if err != nil {
-		slog.Error(
-			"failed to parse issueDate as a timestamp",
-			slog.String("issueDate", crResp.IssueDate),
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	tc.authLoaded = true
-
-	err = tc.saveTelemetryAuth()
-	if err != nil {
-		slog.Error(
-			"failed to save TelemetryAuth",
-			slog.String("err", err.Error()),
-		)
-		return
-	}
-
-	slog.Debug(
-		"successfully registered as client",
-		slog.Int64("clientId", tc.auth.ClientId),
-	)
-
-	return nil
 }
 
 func (tc *TelemetryClient) Generate(telemetry types.TelemetryType, content *types.TelemetryBlob, tags types.Tags) error {
